requirements.txt
absl-py==2.1.0
acvl_utils==0.2
aiofiles==23.2.1
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
alembic==1.13.3
annotated-types==0.7.0
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
azure-ai-documentintelligence==1.0.0
azure-cognitiveservices-speech==1.38.0
azure-core==1.32.0
babel==2.16.0
backoff==2.2.1
batchgenerators==0.25
batchgeneratorsv2==0.2.1
bcrypt==4.2.1
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.8.2
blis==1.0.1
branca==0.8.1
build==1.2.2.post1
cachetools==5.4.0
catalogue==2.0.10
catboost==1.2.7
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.6
chromadb==0.6.3
click==8.1.7
cloudpathlib==0.19.0
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.8.2
comm==0.2.2
confection==0.1.5
connected-components-3d==3.18.0
contourpy==1.2.1
cramjam==2.8.3
cycler==0.12.1
cymem==2.0.8
dataclasses-json==0.6.7
datasets==3.0.0
debugpy==1.8.5
decorator==4.4.2
defusedxml==0.7.1
Deprecated==1.2.14
dicom2nifti==2.4.11
dill==0.3.8
distlib==0.3.8
distro==1.9.0
durationpy==0.9
dynamic_network_architectures==0.3.1
einops==0.8.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
et_xmlfile==2.0.0
executing==2.0.1
fastapi==0.112.2
fastjsonschema==2.20.0
fastparquet==2024.5.0
ffmpeg==1.4
ffmpy==0.4.0
fft-conv-pytorch==1.2.0
filelock==3.15.4
filetype==1.2.0
Flask==3.0.3
Flask-Cors==5.0.0
flatbuffers==24.3.25
folium==0.17.0
fonttools==4.53.1
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.6.1
future==1.0.0
fuzzywuzzy==0.18.0
geopandas==1.0.1
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.15
google-api-core==2.19.1
google-api-python-client==2.138.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-genai==0.5.0
google-generativeai==0.8.4
googleapis-common-protos==1.63.2
gradio==4.42.0
gradio_client==1.3.0
graphviz==0.20.3
greenlet==3.1.1
grpcio==1.65.1
grpcio-status==1.62.2
gryds==0.0.9
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.4
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.24.6
humanfriendly==10.0
humanize==4.10.0
idna==3.7
imagecodecs==2024.6.1
imageio==2.34.2
imageio-ffmpeg==0.5.1
imbalanced-learn==0.13.0
imblearn==0.0
importlib_metadata==8.5.0
importlib_resources==6.4.4
ipykernel==6.29.5
ipython==8.26.0
isodate==0.7.2
isoduration==20.11.0
itsdangerous==2.2.0
jax==0.4.30
jaxlib==0.4.30
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
json5==0.9.25
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kaggle==1.6.17
kiwisolver==1.4.5
kubernetes==31.0.0
langchain==0.3.15
langchain-community==0.3.15
langchain-core==0.3.31
langchain-google-genai==2.0.9
langchain-openai==0.1.20
langchain-text-splitters==0.3.5
langcodes==3.4.1
langsmith==0.2.11
language_data==1.2.0
lazy_loader==0.4
lightgbm==4.5.0
linecache2==1.0.0
lxml==5.3.0
Mako==1.3.5
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.25.1
matplotlib==3.9.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mediapipe==0.10.14
mistune==3.0.2
ml-dtypes==0.4.0
mmh3==5.0.1
mne==1.8.0
monotonic==1.6
MouseInfo==0.1.3
moviepy==1.0.3
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
murmurhash==1.0.10
mypy-extensions==1.0.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
nibabel==5.2.1
nilearn==0.10.4
nltk==3.9.1
nnunetv2==2.5.1
notebook==7.2.2
notebook_shim==0.2.4
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.20.1
openai==1.37.1
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
openpyxl==3.1.5
opentelemetry-api==1.29.0
opentelemetry-exporter-otlp-proto-common==1.29.0
opentelemetry-exporter-otlp-proto-grpc==1.29.0
opentelemetry-instrumentation==0.50b0
opentelemetry-instrumentation-asgi==0.50b0
opentelemetry-instrumentation-fastapi==0.50b0
opentelemetry-proto==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
opentelemetry-util-http==0.50b0
opt-einsum==3.3.0
optuna==4.0.0
orjson==3.10.6
osmium==4.0.2
osmnx==2.0.0
overrides==7.7.0
packaging==24.1
pandas==2.1.4
pandocfilters==1.5.1
parso==0.8.4
pillow==10.4.0
pipenv==2024.0.1
platformdirs==4.2.2
plotly==5.24.1
pooch==1.8.2
posthog==3.8.4
preshed==3.0.9
proglog==0.1.10
prometheus_client==0.21.0
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==5.29.3
psutil==6.0.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==17.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyAutoGUI==0.9.54
pycparser==2.22
pydantic==2.8.2
pydantic-settings==2.7.1
pydantic_core==2.20.1
pydicom==2.4.4
pydub==0.25.1
pyEDFlib==0.1.38
PyGetWindow==0.0.9
Pygments==2.18.0
PyMsgBox==1.0.9
pyogrio==0.10.0
pyparsing==3.1.2
pyperclip==1.9.0
PyPika==0.48.9
pyproj==3.7.0
pyproject_hooks==1.2.0
pyreadline3==3.5.4
PyRect==0.2.0
PyScreeze==1.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-gdcm==3.0.24.1
python-json-logger==2.0.7
python-multipart==0.0.9
python-slugify==8.0.4
pytube==15.0.0
pytubefix==6.7.0
pytweening==1.2.0
pytz==2024.1
pywin32==306
pywinpty==2.0.14
PyYAML==6.0.1
pyzmq==26.1.0
RapidFuzz==3.10.0
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.21.0
rsa==4.9
ruff==0.6.3
safetensors==0.4.5
scikit-image==0.24.0
scikit-learn==1.5.1
scipy==1.14.0
seaborn==0.13.2
semantic-version==2.10.0
Send2Trash==1.8.3
sentence-transformers==3.2.1
shapely==2.0.6
shellingham==1.5.4
SimpleITK==2.4.0
six==1.16.0
sklearn-compat==0.1.3
slack_bolt==1.22.0
slack_sdk==3.34.0
smart-open==7.0.5
smmap==5.0.1
sniffio==1.3.1
sounddevice==0.4.7
soupsieve==2.6
spacy==3.8.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.35
srsly==2.4.8
stack-data==0.6.3
starlette==0.38.4
subprocess.run==0.0.8
sympy==1.13.2
tenacity==8.5.0
terminado==0.18.1
text-unidecode==1.3
thinc==8.3.2
thop==0.1.1.post2209072238
threadpoolctl==3.5.0
tifffile==2024.8.10
tiktoken==0.7.0
tinycss2==1.4.0
tokenizers==0.20.1
tomlkit==0.12.0
torch==2.4.1
torchaudio==2.4.0
torchio==0.19.9
torchvision==0.19.1
tornado==6.4.1
# Editable install with no version control (totalsegmri==0.0.1)
-e c:\users\itsoh\desktop\lumbar segmentation\totalsegmentator-mri
tqdm==4.66.5
traceback2==1.4.0
traitlets==5.14.3
transformers==4.45.2
typer==0.12.5
types-python-dateutil==2.9.0.20241003
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
ultralytics==8.2.99
ultralytics-thop==2.0.8
unittest2==1.1.0
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.2.2
uvicorn==0.30.6
virtualenv==20.26.3
wasabi==1.1.3
watchfiles==1.0.4
wcwidth==0.2.13
weasel==0.4.1
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==14.2
Werkzeug==3.0.4
wrapt==1.16.0
xgboost==2.1.1
xxhash==3.5.0
xyzservices==2024.9.0
yacs==0.1.8
yarl==1.11.1
zipp==3.21.0